Improve updating CKAN resources
Each CKAN data set might have one or more resources.  Previously, the
code disregarded any existing resources and recreated them with a new
ID.  This broke bookmarking resource URLs and following links to
resources if when the shared data harvester runs in between generating
links and requesting them.

This commit retains the IDs, and consequently URLs, of resources, but it
does so by removing them briefly before recreating them with their
previous IDs.  Calling package_patch() instead of package_update() might
help with this, but when I tried this the CKAN client disliked the
StringIO objects.  It's also possible that package_patch() would handle
fields removed from the metadata badly, by leaving them intact instead
of removing them, so we'd need to think carefully about the logic of
this approach.